docker/requirements.txt
boto3==1.34.0
opencv-python-headless==4.8.1.78
ultralytics==8.0.196
Pillow==10.1.0
numpy==1.24.3